requirements.txt
asgiref==3.6.0
Django==4.1.6
mysqlclient==2.1.1
Pillow==9.4.0
sqlparse==0.4.3
tzdata==2022.7
django-crispy-forms==1.14.0
crispy-bootstrap5==0.6
django-allauth==0.50.0
django-cleanup==6.0.0
imageio==2.25.0
pygifsicle==1.0.7
django_q===1.3.9
django-user-sessions==2.0.0
#django_filter==22.1
#django_bootstrap_datepicker_plus==5.0.3
# django-preventconcurrentlogins==0.8.2